Bug 1216460 - [1.1] Refactor data types, fix logs and prevent harmful type promotions in SourceBuffer eviction handling. r=jya

+++ dom/media/mediasource/SourceBufferResource.h
@@ -113,7 +113,7 @@ public:
   }
   // Remove data from resource if it holds more than the threshold
   // number of bytes. Returns amount evicted.
-  uint32_t EvictData(uint64_t aPlaybackOffset, uint32_t aThreshold,
+  uint32_t EvictData(uint64_t aPlaybackOffset, int64_t aThreshold,
                      ErrorResult& aRv);
 
   // Remove data from resource before the given offset.
